profile.cpp
I'm not sure if it's correct, though. I thought it could be helpful, in reality terms, to end the program if the user's age is under 18. I googled it, still pretty unsure of that I added. I'd be much obliged if you could provide me the right answer, with a bit of an explanation...

Thank you in advance :)